gcode_parse.h: raise S word size.
Some M-codes apparently need this. Perhaps it could be avoided,
but it costs just 8 bytes binary size and 2 bytes RAM, so not
really a reason to make a headache.

+++ gcode_parse.h
@@ -47,7 +47,7 @@ typedef struct {
 	uint8_t						M;				///< M command number
 	TARGET						target;		///< target position: X, Y, Z, E and F
 
-	int16_t						S;				///< S word (various uses)
+	int32_t						S;				///< S word (various uses)
 	uint16_t					P;				///< P word (various uses)
 
 	uint8_t						T;				///< T word (tool index)
